week6/requirements.txt

albumentations
einops
huggingface-hub
imageio
librosa
matplotlib
numpy
opencv-python
pillow
pytorch-lightning
PyYAML
scikit-image
scikit-learn
scipy
seaborn
tensorboard
torch
torchaudio
torchmetrics
torchsummary
torchvision
tqdm
transformers
wandb

